requirements.txt
fastapi
uvicorn
slack_sdk
python-dotenv
motor
httpx
requests